+++ database/README.md
@@ -1,74 +1,72 @@
 # Travel Planner App - Database Setup
 
-PostgreSQL credentials:
+PostgreSQL Cloud Database Credentials:
 ```bash
-port: 9010
+port: 5432
 username: cs467
-password: cs467
+password: LIgMhiHaoZi4oXO9G19GtpwisBumHwCS
+host: dpg-csp82ol6l47c73epvjg0-a.virginia-postgres.render.com
+database: travel_planner_db_hoi8
 ```
 
 ## Prerequisites
 
-- Install PostgreSQL on your machine.
-https://www.postgresql.org/download/windows/
-- Install pgAdmin (optional, for GUI access to the database).
+- Install pgAdmin (optional) for a GUI tool to manage the database.
 https://www.postgresql.org/ftp/pgadmin/pgadmin4/v8.12/windows/
 - Ensure Node.js and npm are installed.
+- Install dotenv package if not already done for loading environment variables:
+```bash
+npm install dotenv
+```
 
-## Database Setup
+## .env File Configuration
+- You will need to create a ".env" file in the backend folder of the project with the following content:
+```bash
+DB_USER="cs467"
+DB_HOST="dpg-csp82ol6l47c73epvjg0-a.virginia-postgres.render.com"
+DB_DATABASE="travel_planner_db_hoi8"
+DB_PASSWORD="LIgMhiHaoZi4oXO9G19GtpwisBumHwCS"
+DB_PORT=5432
+```
+Note: The ".env" file should also be added to the .gitignore in the backend folder to prevent pushing it on GIT.
 
-1. **Start PostgreSQL Service**  
-   Make sure your PostgreSQL service is running.
+## Database Connection Setup in pgAdmin (Optional but Recommended)
 
-2. **Open PostgreSQL CLI**  
-   Open a terminal and connect to PostgreSQL with the following command:
-   ```bash
-   psql -U postgres
+1. Open pgAdmin and create a new server connection:
 
-3. Run these commands to set up the database and user for the project:
-    ```bash
-    CREATE DATABASE travel_planner_db;
-    CREATE USER cs467 WITH PASSWORD 'your_password';
-    ALTER ROLE cs467 SET client_encoding TO 'utf8';
-    ALTER ROLE cs467 SET timezone TO 'UTC';
-    GRANT ALL PRIVILEGES ON DATABASE travel_planner_db TO cs467;
-    ```
+    - Name: Render Travel Planner
+    - Host name/address: dpg-csp82ol6l47c73epvjg0-a.virginia-postgres.render.com
+    - Port: 5432
+    - Username: cs467
+    - Password: LIgMhiHaoZi4oXO9G19GtpwisBumHwCS
+    - Database: travel_planner_db_hoi8
+    - SSL Mode: If required, set to Require or Verify CA.
 
-4. Run Schema and Seed Files:
+2. Test the connection to ensure pgAdmin connects successfully.
 
-    Exit from the psql prompt with '\q' and reconnect to the new database:
-    ```bash
-    psql -U cs467 -d travel_planner_db
-    ```
-    Then run the schema.sql and seed.sql files.
-    
-    If it's easier, just copy and paste the schema.sql and seed.sql files into pgAdmin. That worked best for me as directly manipulating the database in the terminal was a bit tricky.
+## Running and Testing the Local server
 
-    Verify that the tables have been populated with:
-    ```bash
-    SELECT * FROM Users;
-    SELECT * FROM Trips;
-    SELECT * FROM Experiences;
-    SELECT * FROM Ratings;
-    ```
+1) Start the local server:
 
-## Backend Setup for database
-
-1. Open 'backend/config/db.js' and make sure the user, password, database, and port match your local PostgreSQL setup.
-
-2. Start the Server
-
+First navigate to the backend folder of the project and run:
 ```bash
 node server.js
 ```
 
-3. Test API Endpoints
+Make sure that the server starts without and errors. You should see a message like:
+```bash
+Server is running on port 5000
+```
 
-Either use a tool like Postman or directly in the browser.
+2) Test API Endpoints
 
-For example:
-- GET all users: http://localhost:5000/api/users
+- Use Postman or in the web browser directly to test API endpoints locally.
+- Example: To get all users, make a GET request to:
+```bash
+http://localhost:5000/api/users
+```
 
 ## Notes
 
-This should help everyone get the database up and running locally, but I will work to improve this setup as well as potentially migrating to an online database if that is easier and I can find a free host.
+- Ensure your .env file is set up correctly before starting the server.
+- If you encounter connection issues, double-check the credentials and SSL settings in pgAdmin and your .env file.
